Pubsub query (#246)
* chore: bump zenoh version

* Retrigger CI

* chore: update branch

* chore: rebase onto dev/1.0.0

* fix: fix
